Add output element properties update
Output element such as Streaming and Storing can be updated
after their creation. But in order to take into account the
changes, the pipeline must be stopped and then launched.
This is not a dynamic update.
Note that a change of feed type (e.g. audio_only --> video_only)
will not change the type of the element choosen at its creation.

Remove set_speaker_volume method since a click on the speaker
button just mutes the audio output.