[FIX] Video freezing when loading second session
When a session was loaded and then a second one later,
the video feed was freezing (the pipeline wasn't switching
back to preview mode). This behavior is fixed and sessions
can be loaded as many as it needed without restarting the
software.

Moreover, the load confirmation dialog is not displayed
anymore when a session is loaded in the where the
software is launched without any input sources
AND output sinks set.