Add streaming server auto-reconnect
When an Icecast server is down, the pipeline is notified and if nothing
is done it will most likely crash. Now when one loses the connection to
an Icecast instance, the pipeline will deconnect the faulty branch.
Then it actively tries to reconnect to the server and will do so until
the pipeline is stopped.