Add checkbox to mute loudspeakers/headphones audio output
This checkbox has been added as a mirror of the one used to mute
audio input source. It was confusing to have a checkbox in the menu
for the input and a ToolButton in the control bar to mute the output.
Nevertheless, this new checkbox is here only for visual purpose and
thus is set to unsensitive. The loudspeaker output is still requested
by the ToolButton in the control bar.