Ask for confirmation on session loading when main pipeline is playing
This remove unnecessary complexity. It now asks user for a confirmation
whenever the placeholder pipeline is not playing. This happends once a
video or an audio input is selected.
This also fixes a bug occurring when one tried to load a session onto
an existing one. Bug caused by an AttributeError in the Pipeline object.